# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinPilot",
    "description": "Authenticated financial workspaces with deterministic calculations.",
    "version": "2.0"
  },
  "paths": {
...
    "/api/overview": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Get the money overview",
        "description": "Returns the household money overview.",
        "operationId": "overview_api_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
    "/api/liquidity": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Get liquidity tiers",
        "description": "Groups balances into tiers by how quickly the money can be used.",
        "operationId": "liquidity_api_liquidity_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
    "/api/coverage": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Estimate deposit insurance coverage",
        "description": "Estimates deposit insurance coverage across accounts and ownership categories.",
        "operationId": "coverage_api_coverage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
    "/api/bills": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "List upcoming bills",
        "description": "Lists bill occurrences due within the next `days` days.",
        "operationId": "bills_api_bills_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 366,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/forecast": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Forecast cash",
        "description": "Projects dated balances for one account, or all planning accounts, over the next `days` days.",
        "operationId": "forecast_api_forecast_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 366,
              "minimum": 1,
              "default": 45,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/allowance": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Calculate a spending allowance",
        "description": "Calculates how much can be spent over the next `days` days while still covering obligations and protected cash.",
        "operationId": "allowance_api_allowance_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 366,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/buffer": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Calculate an operating buffer",
        "description": "Calculates the operating buffer for an account.",
        "operationId": "buffer_api_buffer_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/connections": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "List account connections",
        "description": "Lists the connections between accounts in the workspace.",
        "operationId": "connections_api_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
    "/api/recurring/activity": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "List recurring rule activity",
        "description": "Lists recurring rule runs within `horizon_days` and whether each will run, is paused, skipped or lacks authorization.",
        "operationId": "recurring_activity_api_recurring_activity_get",
        "parameters": [
          {
            "name": "horizon_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 366,
              "minimum": 1,
              "default": 60,
              "title": "Horizon Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/plan": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Get the paycheck plan",
        "description": "Returns the paycheck allocation plan for a month, the current month by default.",
        "operationId": "plan_api_plan_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2200,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/policies": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Get rule automation status",
        "description": "Returns the status of recurring rules and their authorization.",
        "operationId": "policies_api_policies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      },
      "post": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Create a fixed rule (legacy)",
        "description": "Creates a recurring rule for an older form; new clients use `POST /api/manage/{kind}`. A saved rule has no payment authority.",
        "operationId": "create_policy_api_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/policies/{policy_id}/authorize": {
      "post": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Authorize a rule for simulation",
        "description": "Gives a rule simulation-only payment authority with a per-run cap.",
        "operationId": "authorize_policy_api_policies__policy_id__authorize_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          }
        }
      }
    },
    "/api/policies/{policy_id}/pause": {
      "post": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Pause or resume a rule",
        "description": "Pauses a rule, or resumes it when `paused` is false.",
        "operationId": "pause_policy_api_policies__policy_id__pause_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Id"
            }
          },
          {
            "name": "paused",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Paused"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          }
        }
      }
    },
    "/api/policies/{policy_id}/skip-next": {
      "post": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Skip a rule's next run",
        "description": "Skips the next run of a rule.",
        "operationId": "skip_next_policy_api_policies__policy_id__skip_next_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          }
        }
      }
    },
    "/api/bills/{bill_id}/pay-once": {
      "post": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Draft a one-time bill payment",
        "description": "Prepares a payment draft for a bill occurrence. Nothing is submitted.",
        "operationId": "pay_bill_once_api_bills__bill_id__pay_once_post",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bill Id"
            }
          },
          {
            "name": "occurrence_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurrence Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The If-Match header does not contain a workspace revision."
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          }
        }
      }
    },
    "/api/debt/compare": {
      "post": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Compare debt payoff strategies",
        "description": "Compares debt payoff strategies, optionally with an extra monthly payment sent in the JSON body.",
        "operationId": "debt_compare_api_debt_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebtCompareInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
...
    "/api/tax/profile": {
      "get": {
        "tags": [
          "Financial workspace"
        ],
        "summary": "Get tax assumptions",
        "description": "Returns the workspace tax assumptions.",
        "operationId": "tax_profile_api_tax_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorizationIn": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AuthorizationMode",
            "default": "standing"
          },
          "per_run_cap": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Per Run Cap"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "per_run_cap"
        ],
        "title": "AuthorizationIn"
      },
...
      "DebtCompareInput": {
        "properties": {
          "extra_payment": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Payment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DebtCompareInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}